add spanProcessor
