+++ pkg/trace/traceutil/span_processor.go
@@ -1,0 +1,14 @@
+// Unless explicitly stated otherwise all files in this repository are licensed
+// under the Apache License Version 2.0.
+// This product includes software developed at Datadog (https://www.datadoghq.com/).
+// Copyright 2016-present Datadog, Inc.
+
+package traceutil
+
+import (
+	"github.com/DataDog/datadog-agent/pkg/trace/pb"
+)
+
+type SpanProcessor interface {
+	Process(tags map[string]string, span *pb.Span)
+}
